agents/backend/onyx/server/features/heygen_ai/requirements-ml.txt
# Machine Learning Module Requirements for HeyGen AI
# Core deep learning libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and language models
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.12.0
accelerate>=0.20.0

# Diffusion models
diffusers>=0.18.0
xformers>=0.0.20

# Gradio for interfaces
gradio>=3.35.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
structlog>=23.0.0
loguru>=0.7.0

# Experiment tracking
tensorboard>=2.13.0
wandb>=0.15.0

# Performance optimization
psutil>=5.9.0
pyinstrument>=4.0.0
memory-profiler>=0.61.0

# Optional: GPU optimization
# nvidia-ml-py3>=11.0.0  # For GPU monitoring
# apex  # For mixed precision training (install from source)

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter for development
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional utilities
pillow>=9.5.0  # For image processing
matplotlib>=3.7.0  # For plotting
seaborn>=0.12.0  # For statistical plots
plotly>=5.15.0  # For interactive plots

# Security and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async support
aiofiles>=23.0.0
asyncio-mqtt>=0.11.0

# Database support (optional)
# sqlalchemy>=2.0.0
# asyncpg>=0.28.0
# motor>=3.1.0
# redis>=4.6.0

# Web framework (if needed)
# fastapi>=0.100.0
# uvicorn>=0.23.0
